// migrations/snapshots/973db2799b29f4a10cad2db8b70617b072ebf997090f8737b011521588685b12/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly InventoryBatch: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['output'];
      readonly batchNumber: CodecTypes['pg/text@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly unitCost: CodecTypes['pg/numeric@1']['output'];
      readonly totalCost: CodecTypes['pg/numeric@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly InventoryBatchStock: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly batchId: CodecTypes['pg/int4@1']['output'];
      readonly locationId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly reservedQuantity: CodecTypes['pg/numeric@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly unit: CodecTypes['pg/text@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly InventoryLocation: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly InventoryTransfer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly batchId: CodecTypes['pg/int4@1']['output'];
      readonly fromLocationId: CodecTypes['pg/int4@1']['output'];
      readonly toLocationId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Product: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly sellingPrice: CodecTypes['pg/numeric@1']['output'];
      readonly finishedProductItemId: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly ProductCostHistory: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly hpp: CodecTypes['pg/numeric@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Production: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly recipeId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly totalHpp: CodecTypes['pg/numeric@1']['output'];
      readonly unitHpp: CodecTypes['pg/numeric@1']['output'];
      readonly batchId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly ProductRecipe: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly version: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly RecipeItem: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly recipeId: CodecTypes['pg/int4@1']['output'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Restock: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['output'];
      readonly batchId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly totalCost: CodecTypes['pg/numeric@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly InventoryBatch: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['input'];
      readonly batchNumber: CodecTypes['pg/text@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly unitCost: CodecTypes['pg/numeric@1']['input'];
      readonly totalCost: CodecTypes['pg/numeric@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly InventoryBatchStock: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly batchId: CodecTypes['pg/int4@1']['input'];
      readonly locationId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly reservedQuantity: CodecTypes['pg/numeric@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly InventoryItem: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly unit: CodecTypes['pg/text@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly InventoryLocation: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly InventoryTransfer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly batchId: CodecTypes['pg/int4@1']['input'];
      readonly fromLocationId: CodecTypes['pg/int4@1']['input'];
      readonly toLocationId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Product: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly sellingPrice: CodecTypes['pg/numeric@1']['input'];
      readonly finishedProductItemId: CodecTypes['pg/int4@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly ProductCostHistory: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly hpp: CodecTypes['pg/numeric@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Production: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly recipeId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly totalHpp: CodecTypes['pg/numeric@1']['input'];
      readonly unitHpp: CodecTypes['pg/numeric@1']['input'];
      readonly batchId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly ProductRecipe: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly version: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly RecipeItem: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly recipeId: CodecTypes['pg/int4@1']['input'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Restock: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly inventoryItemId: CodecTypes['pg/int4@1']['input'];
      readonly batchId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly totalCost: CodecTypes['pg/numeric@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;